(Discord) Send Twitch relay messages to designated channel using #channel prefix in Discord
If the bot is in ValZarGaming's or DAAthren's channel, sending the message `#valzargaming hello` or `#daathren hello` in the proper Discord channel will send the message `[DISCORD] Username: hello` to the named channel. If the channel does not exist the message will be sent to the last known channel or the first channel in the channels array if none has been used yet. Additionally, sending a message via Discord without specifying a channel will always go to the channel that last had a message sent to it, or the first channel in the channels list if no message has been sent yet.